Fixes: 0001266: Conditional Grouping
This problem only happens in group by group mode when the condition is based in a multiple options (with or without comments) question type that is presented in the third page (second page with questions) and only when the survey is active.

This is because the default values for questions is NULL and isset($_SESSION[$fieldname]) evaluates to false even if $_SESSION[$fieldname] is set, but to a NULL value.

This don't happend if the multiple option question is presented in the firs page with questions, because save.php sets $_SESSION[$fieldname] to "".

git-svn-id: file:///Users/Shitiz/Downloads/lssvn/source/limesurvey@3344 b72ed6b6-b9f8-46b5-92b4-906544132732