Spanish language updated.
This translation was done directly from new en.po, using but revising the 1.5* old spanish translation strings one by one.

There were corrected A LOT of msgstr's, and also I changed the translation of "answer/answers" when defining the different answers (options in fact) of a question as "opción/opciones", which is a much more accurate than "respuesta" in this context (at least in an spanish parlant environment :-) ).

git-svn-id: file:///Users/Shitiz/Downloads/lssvn/source/limesurvey@4328 b72ed6b6-b9f8-46b5-92b4-906544132732