Fixed Bug - Conditions to Relevance was AND-ing instead of OR-ing scenarios Fixed Bug - Conditions to Relevance was not detecting No Answer - switched to using is_empty(SGQA) Fixed Bug - Hidden status wasn't always processed correctly Fixed Bug - Multiple Choice and List questions were appearing as Unanswered even when partially answered Fixed Bug - A value of 0 was considered unanswered
Fixed Bug - PHP and JavaScript handle 0, '0', and '' differently.  Changed EM to keep them as aligned as possible. Despite this, there are still some cases where PHP and JavaScript will disagree:
(1) 0 == '' is true in JavaScript and false in PHP
(2) 0 != '' is false in JavaScript and true in PHP
(3) '' < 10 is true in JavaScript and false in PHP
So, if there are conditions like X < 16, then the author must use (!is_empty(X) and (X < 16)), otherwise a blank value will always meet the condition of X < 16



git-svn-id: file:///Users/Shitiz/Downloads/lssvn/source/limesurvey_dev_tms@11657 b72ed6b6-b9f8-46b5-92b4-906544132732