Fixed issue 06484: EM sessions misbehaving
This line should fix the majority of the session troubles with EM because setting the survey forces EM to reload if the surveyid has changed. I still don't think this is quite the right approach, however, because the original survey is reset. I think we should work to move all frontend session variable in to different session arrays (each with ID survey_######) so that two or more surveys can be taken concurrently without issue.