Allow 0 (as op to. null) value in POST/GET/COOKIES.
returnGlobal() returned NULL when the $stringname argument was an
existing GET or POST or COOKIE variable containing the '0' string.

This broke on PostgreSQL, for instance when saving the email settings
with emailsmtpdebug off/disabled ($_POST['emailsmtpdebug'] is '0') :
    application/controllers/admin/globalsettings.php does :
        if (returnGlobal('emailsmtpdebug')) setGlobalSetting('emailsmtpdebug', sanitize_int(returnGlobal('emailsmtpdebug')));
which therefore tries to save a null value in the NOT NULL
prefix_settings_global's stg_value column.

We will return "0" instead of null in that case.