New feature #8214: DateTimepicker for date/time question
Dev: Also some cleanup in date.js
Dev: You can now not only pick a date but
Dev: also a time or a combination or just month and
Dev: year in the date question.
Dev: Question to the pros: The timepicker is not yet i8n.
Dev: How can I use gT() in a .js?