Fixed issue #9026: If empty survey, export 1 empty record to avoid blank results.
A blank page was returned when exporting empty surveys, because no content was given to output. 
Headers were not given neither (as contrary to the expected), as headers are written before the first record. As no records were given, no headers were written.
By providing an empty record, headers are written and output matches expectations.