Update AuthLDAP.php
The existing code assumes that the user will enter a mail attribute in lower case.  This will often be the 'mail' attribute which causes no problems.  Where the desired email attribute is in U/L case as in for example; alternateEMail, the user creation will fail because the ldap_get_entries array only contains lower case values.