New feature: Allow test & basic feedback of the email functionality.
New feature: Allow test & basic feedback of the email functionality.

Dev: This change affects the UI of two of the tabs - "bounce settings" and "email settings"
Dev: the UI changes are just to keep everything consistent.
Dev: The main change adds a "Test" button to the "email settings" tab which allows
Dev: the user to test the *saved* configuration options. Feedback is returned and displayed
Dev: in a flash message.