bug fix 14268
LDAP import. Checkbox "filterduplicatetoken" and "filterblankemail" send values != '1'. 
there is no need to check the value of the variable, it is enough that it exists in the POST request.
without these changes is not working filter "Filter Duplicates", this fix solve problem in https://bugs.limesurvey.org/view.php?id=14268