Fixed issue #16483: Adding condition triggers bulk update for complete database
Function updated to get the survey ID from the question if surveyId is null and a qid is specified.

UpgradeConditionsToRelevance receives two params: surveyId and qid. If the qid is specified, it should only update that question (as per the function doc). But, if surveyId is null, it loops trough all the surveys, which only makes sense if no qid is specified.