Fixed issue #16566: We cannot create multiple choice questions
Before issue #16469, QuestionEditorController tried to find the questions using 'findByPk', which automatically casts the value to int. Now, 'find' is used because the survey ID, as the search is done by multiple criterias.
When saving a new question, 'qid' is an empty string, and the resulting SQL fails in Postgres because of the data type.
Solved by casting 'qid' to int.
This also helps with issue #16563.