Fixed issue #16549: Error message is shown while trying to delete and reupload file using File upload
Removed the iframe after receiving the result.
Also fixed the order of _getResponse and form.submit() calls, which doesn't seem to affect this case but was wrong anyway: _getResponse only adds the load event for the iframe, and it was being called after form.submit(). The event should be defined before submitting the form.