Fixed issue #16194: file upload disables other clicks.
After loading the upload frame inside the modal, 'relevance:on' event handlers are lost.
Calling triggerEmRelevance() when closing the modal fixes the issue.

Removes the cache for the currentUploadHandler. This wasn't working for multiple upload question in the same as it was caching it on a window level variable. Removed the cache for simplifiying.

Contains also fix for #16549: Error message is shown while trying to delete and reupload file using File upload