Fixed issue #16373: Bootstrap Buttons not aligned when filtered
The element that is normally hidden by relevance (with ID starting with 'javatbd') is inside a "bootstrap-buttons-div" element in the case of bootstrap buttons. The "bootstrap-buttons-div" element has a 'fixed' size (ex: col-12). So when hidding, the space remains blank even if the button is hidden.

Changed 'hideQuestionWithRelevanceSubQuestion' (template-core.js) to also hide the "bootstrap-buttons-div".