Fixed issue #16634: Timer not getting reset after completion of the survey (running multiple surveys on the same browser)
Each timer registers itself in a local storage item identified by the Survey ID.
This is used by the reset function to know which timers belong to the survey and are to be reset.
The reset method was replaced by a function (in survey.js instead of timer.js) that reset all timers for a given survey.
Also reviewed the reset method for preview mode.