Fixed issue #16633: Error message when changing the data policy message language unless I check the box (#1574)
* Fixed issue #16633: Error message when changing the data policy message language unless I check the box

Making a form of its own for the language changer of the first page. This form as only have the language will not trigger server side validation of the policy. This is how it is in LS4

* Revert "Fixed issue #16633: Error message when changing the data policy message language unless I check the box"

This reverts commit a48b2b5ff950d7ab2904575d8feacd072783b198.

* Fixed issue #16633: Error message when changing the data policy message language unless I check the box

Making a form of its own for the language changer of the first page. This form as only have the language will not trigger server side validation of the policy. This is how it is in LS4.

Language changer from menu also updated.