Fixed issue #16644: Audit log not save all actions
There was a mismatch between the default values for "Audit log for this survey" setting. The default value shown on screen was Yes (true), and for most events it was indeed treated as 'true' by default when logging actions. But for tokens and survey settings, it defaulted to 'false' when logging actions.

It was fixed by defaulting to 'true' in all cases, as shown in the screen but expected by the user.