Fixed issue #16560: After downloading tokens for the first time, need to refresh for a second download (#1613)
When "save buttons" are clicked, they are disabled, the spinner is shown, and the form is submitted. It is expected that the submit ends in a redirect.
The export button was identified as a "save button", but ended in a file download.
In this case, since a file is donwloaded, the page remains in the same state, with the spinner displayed and the button disabled.

It was changed to be handled as an "export button".