Fixed issue #16670: Extended theme shows all options in one page
When the theme config is copied, the engine node is removed (apparently it caused issues before).
Since the optionspage config is inside the engine node, it's removed from the extended theme config, and then the options.twig is used.
One option is to move the optionspage out of the engine node, so it's not removed.
Other option is to remove the engine node contents EXCEPT optionspage, instead of the whole engine node. This option seems easier and it was the one implemented.