Fixed issue #16776  question index marks pages with fully answered questions as unanswered (#1652)
* Fixed issue #16776: Question index marks pages with fully answered questions as unanswered

For Numeric Array question types with Checkbox layout, if is enough for mandatory, we flag it as answered. If not, we flag it as anunserwed.

* Fixed issue #16776: Question index marks pages with fully answered questions as unanswered

For Numeric Array question types with Checkbox layout, if is enough for mandatory, we flag it as answered. If not, we flag it as anunserwed.

* Fixed issue #16776: Question index marks pages with fully answered questions as unanswered

Dummy change to trigger travis