Fixed issue #16898: Pop Up Editor does not work inside the Question Editor
There were issues with the triggering of the popup editor, both for answers and for subquestions.
These were mainly related to the identification of the target field, and also the values of SID, GID and QID passed to the editor.
Some problems originated in the twigs, but the updateRowProperties() method (questionEditor.js) was also causing problems.