Register assets
How to reproduce:

disable the Limesurvey XSS filter at
Global settings -> Security -> “Filter HTML for XSS” = “No”.

Create a "Short Free Text" question.

Add a script to the question. Example: `console.log (" I'm here ")`

Preview the issue.

Check to see if you’re here in the developer tools console of your browser.

The correction:
The method needs to be refactored but the change resolves the bug.