Fixed issue #15819: Missing localization for group_name in add_group, set_group_properties, import_group
'import_group' was working fine.
'add_group' was updated as to save to the QuestionGroupL10n model.

'set_group_properties' was updated as to be able to receive an array on 'questiongroupl10ns' param.
The method is backward compatible. If an 'questiongroupl10ns' is not given, then it is composed from the old name, description and language parameters.

Adding test for 'add_group' y 'set_group_properties'.