Dev Add new CI pipeline for 3.x-LTS (#1952)
The new pipeline will test the code on all branches that will have
this workflow file. The tests will ran with php 5.6. and 7.4. .
Both jobs will run, even if some fail. The results of the tests can
be found in the action tab, on a repo page.

Changes made to php code were necessary, and were discussed when first
version of CI pipeline (for master branch) was developed.