Fixed issue #17982: Malformed DSN when using MySQL socket and DB name during installation
How to reproduce:

* create a database
* create an user@'%' identified by a password
* grant all privileges to user@'%' to that database
* install LimeSurvey and fill MySQL configuration
* remove "localhost" as MySQL server and set a valid socket like "/var/run/mysqld/mysqld.sock"
* set the database name normally
* save

What it happens:

«Database doesn't exist!»

What should happen if the database exists and if the MySQL user can see the database:

Normal installation.

Cause of the problem:

It seems that when the user selects a valid socket AND fills the database name,
then DSN is wrongly formatted, without the ';' separator between the socket path
and the database name. This problem exists since the feature was originally introduced
here:

* 4f97fb02e96c2526fab1ef6db5e9cc73eebc1943

The fix consists in assuring that the DSN string ends with the ';' delimiter
also in the case of an user using a socket.

Some users never encountered this problem because they left the database name
blank and put the database name directly into the sock path.

P.S.

The commit includes also a strict string comparison using `===` operator that
is more appropriate when comparing a string with another string.