Update timer.min.js
See: 
https://bugs.limesurvey.org/view.php?id=18784
The timer.js is working as it should! It removes and add the class (bootstrap5) 'd-none' from the warning text.
But the timer.min.js is removing the old (bootstrap<5) class 'hidden' this class is not applied to the text. So the warning stays on 'display:none;' (i.e. d-none)

So timer.min.js needs to be updated!!!