Fixed issue #19769: Duplicate question type images
Dev: When creating or editing a question and selecting a question type, the dialog lists the question types with images. Some question types showed images associated with other question types. I've fixed this by getting the proper images from the official documentation, adding them to the assets and updating the file paths, therefore ensuring that the correct image is displayed for each question type.