Task accessibility update: "Survey list" (#4264)
* Survey list - Changing image role and heading structure

"The following surveys are available" is changed to h2 and added role and label to LS logo

* menu is not keyboard accessible

Changed the div to button so that the menu can also be accessed through keyboard and screen readers.

* Fixing alternative text

The screen reader announces the insufficient alternative text as 'LimeSurvey Survey Software', it should announce the sufficient alternative text as 'Proudly powered by Lime survey' for the lime survey logo image..

* Landmarks are missing

Landmarks are missing on the page, the screen reader should announce the landmarks as follows:
- Banner
- Main
- Content information

* Link is not distinguished

The 'The Online Survey Tool' link is distinguished from the surrounding text solely via color. The 'The Online Survey Tool' link should be easily distinguishable from regular text. Styling options can include underlining or some other distinguishable style.

* Fixing Language dropdown

The option icon are not focusable by using keyboard. the icon should be focusable by using keyboard.

* Addressing PR feedback

1) Alternative text changed to "proudly powered by LimeSurvey"
2) <article> tag changed to <div>

* For old screen readers compatibility, adding roles to HTML5 elements

* Translation fix

---------

Co-authored-by: Carsten Schmitz <carsten.schmitz@limesurvey.org>