Fixed issue: applied a11y to welcome page and text questions (#4291)
* Attached label to Change language

Select element doesn't have an accessible name, ensure select element has an accessible name.

* Fixed heading structure

The screen reader announces the following visual headings as plain text. It should announce the appropriate heading levels as follows:
- Sample LimeSurvey EN/FR/DE/IT/RU as <h1>
- This survey is anonymous. as <h2>

* making tooltip dismissable

The progress tooltip 'You have completed 0% of this survey' is not dismissible using the 'Esc' key on the mouse pointer; it should be dismissible using the 'Esc' key.

* Focus indicator added on language button

The focus indicator is not visible on the button. The focus indicator should be visible on the button.

* Removed static width

The text 'English-English' gets truncated when text spacing properties are applied. The text should not be truncated when applying text spacing properties.

* Changed heading level of question group title

The screen reader announces the visual heading 'Text question' as plain text, it should announce the heading level as 1.

* Title updated globally

The screen reader announces the insufficient page title as 'Sample LimeSurvey EN/FR/DE/IT/RU', it should announce the sufficient page title as 'Text question |Sample LimeSurvey EN/FR/DE/IT/RU'.

Other Occurrences:
This issue has been observed throughout the web application.

* Main landmark added

The screen reader does not announce any landmarks on the page. It should announce at least the 'Main' landmark.

Other Occurrences:
This issue has been observed throughout the web application.

* group and group label attached

The screen reader does not announce the group label with the edit field, such as 'Short free text'. It should announce the group label as '1 S'.

* Descriptive label for navigation button

The screen reader announces the insufficient label as 'Next' and 'Previous', they should announce the sufficient label as 'Next question' and 'Previous question' for the next question and previous question buttons.

* incorrect role removed from the table

The screen reader does not announce the tabular data as a table. It should properly convey the structure as follows:
- Announce the caption as 'Array Texts – This is an array question'
- Identify 5 rows and 5 columns
- Ensure all cells are correctly associated with their respective headers.

* Feedback changes

* Title handled through theme_js_modules

* Role removed from twig file itself

* Dev: rebuild theme files

---------

Co-authored-by: Patrick Teichmann <patrick.teichmann@limesurvey.org>