Fixed issue: applied a11y instruction is not programmatically associated and other accessibility fixes (#4322)
* Attached label to Change language

Select element doesn't have an accessible name, ensure select element has an accessible name.

* Fixed heading structure

The screen reader announces the following visual headings as plain text. It should announce the appropriate heading levels as follows:
- Sample LimeSurvey EN/FR/DE/IT/RU as <h1>
- This survey is anonymous. as <h2>

* making tooltip dismissable

The progress tooltip 'You have completed 0% of this survey' is not dismissible using the 'Esc' key on the mouse pointer; it should be dismissible using the 'Esc' key.

* Focus indicator added on language button

The focus indicator is not visible on the button. The focus indicator should be visible on the button.

* Removed static width

The text 'English-English' gets truncated when text spacing properties are applied. The text should not be truncated when applying text spacing properties.

* Changed heading level of question group title

The screen reader announces the visual heading 'Text question' as plain text, it should announce the heading level as 1.

* Title updated globally

The screen reader announces the insufficient page title as 'Sample LimeSurvey EN/FR/DE/IT/RU', it should announce the sufficient page title as 'Text question |Sample LimeSurvey EN/FR/DE/IT/RU'.

Other Occurrences:
This issue has been observed throughout the web application.

* Main landmark added

The screen reader does not announce any landmarks on the page. It should announce at least the 'Main' landmark.

Other Occurrences:
This issue has been observed throughout the web application.

* group and group label attached

The screen reader does not announce the group label with the edit field, such as 'Short free text'. It should announce the group label as '1 S'.

* Descriptive label for navigation button

The screen reader announces the insufficient label as 'Next' and 'Previous', they should announce the sufficient label as 'Next question' and 'Previous question' for the next question and previous question buttons.

* incorrect role removed from the table

The screen reader does not announce the tabular data as a table. It should properly convey the structure as follows:
- Announce the caption as 'Array Texts – This is an array question'
- Identify 5 rows and 5 columns
- Ensure all cells are correctly associated with their respective headers.

* Feedback changes

* instruction 'Only numbers may be entered in these fields.' is not programmatically associated

The instruction 'Only numbers may be entered in these fields.' is not programmatically associated with the 'Some example subquestion' input field. As a result, it is not announced by the screen reader and should be correctly associated.

* Heading not announced

The screen reader announces the visual heading 'Numeric question' as plain text, it should announce the heading level as 1.

* The screen reader doesn't announce the ' incorrect total count 12', it should announce automatically.

* The screen reader doesn't announce the 'correct total count 10', it should announce automatically.

* When the user navigates to the 'Some example subquestion' slider, the screen reader unnecessarily announces the table; it should not do so

* The screen reader announces the table structure incorrectly as '1 row and 7 columns'; it should announce the correct structure as '5 rows and 7 columns.' Additionally, the table should be accessible using specific table navigation keys in the screen reader.

* When the user navigates to the 'Some example subquestion', the grouping label is not associated with the combo box. The screen reader should announce it as 'Scale A'.

* The screen reader announces the following visual heading as plain text, they should announce the heading level as 2

* When the user navigates to the 'Some example answer option' dragging element, the instruction 'Please select 2 answers' is not announced with it. The screen reader should announce this instruction.

* Handled survey format is "all in one" scenario

* script removed from twig and Improved navigation buttons label

* group role and aria-labelledby removed

* Removed code of different PR

* Role and aria-labelledby reverted