Fix Attribute category being shown in Question settings.
Previously the setting was only unset if the page was displayed in English. Since the category name is translated at this point, it remains in the settings list. This PR solves this issue.

Perhaps there is another way to not translate the category in the first place, but this should work as a quick fix.